Add registry provider to config

+++ Program.cs
@@ -1,45 +1,46 @@
 using System.Runtime.InteropServices;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Win32;
+using RymRss.Configuration;
 using RymRss.Db;
 using RymRss.Models.Options;
 using RymRss.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
-string? dataFolderPath = null;
-string? username = null;
 if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
 {
-    using var key = Registry.LocalMachine.OpenSubKey(@"Software\RymRss");
-    if (key?.GetValue("AppOptions:DataFolder") is string dataFolderValue) dataFolderPath = dataFolderValue;
-    if (key?.GetValue("ScrapeOptions:User") is string userValue) username = userValue;
+    builder.Configuration.AddRegistryKey(Registry.LocalMachine, @"Software\RymRss");
 }
-
+var dataFolderPath = builder.Configuration.GetValue<string?>("AppOptions:DataFolder");
+var username = builder.Configuration.GetValue<string?>("ScrapeOptions:User");
 if (dataFolderPath is not null)
 {
     var settingsFilePath = Path.Join(dataFolderPath, "appsettings.json");
-    if (username is not null)
+    if (username is not null && File.Exists(settingsFilePath))
     {
         var settingsFileText = File.ReadAllText(settingsFilePath);
         settingsFileText = settingsFileText.Replace("{{RYM_USERNAME}}", username);
         File.WriteAllText(settingsFilePath, settingsFileText);
     }
-    builder.Configuration.AddJsonFile(settingsFilePath);
+    builder.Configuration.AddJsonFile(settingsFilePath, true);
 }
-builder.Configuration
-    .AddInMemoryCollection(new Dictionary<string, string?>
-    {
-        ["AppOptions:DataFolder"] = dataFolderPath,
-    });
-var appOptions = builder.Configuration.GetRequiredSection(nameof(AppOptions)).Get<AppOptions>();
+if (dataFolderPath is not null)
+{
+    builder.Configuration
+        .AddInMemoryCollection(new Dictionary<string, string?>
+        {
+            ["AppOptions:DataFolder"] = Environment.ExpandEnvironmentVariables(dataFolderPath),
+        });
+}
+var appOptions = builder.Configuration.GetRequiredSection(nameof(AppOptions)).Get<AppOptions>()!;
 
 builder.Services.AddControllersWithViews();
 builder.Services.Configure<ScrapeOptions>(builder.Configuration.GetRequiredSection(nameof(ScrapeOptions)));
 builder.Services.Configure<AppOptions>(builder.Configuration.GetRequiredSection(nameof(AppOptions)));
 builder.Services.AddDbContext<RymRssContext>(options =>
 {
-    var dbPath = Path.Join(dataFolderPath, "rymrss.db");
+    var dbPath = Path.Join(appOptions.DataFolder, "rymrss.db");
     options.UseSqlite($"Data Source={dbPath}");
 });
 builder.Services.AddHostedService<RymScraper>();
@@ -65,4 +66,4 @@ app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Rym}/{action=Rss}");
 
-app.Run($"http://localhost:{appOptions!.Port}");
+app.Run($"http://localhost:{appOptions.Port}");
